packet-utils: hoist Fletcher LSA checksum shared by OSPF and IS-IS
lsa_checksum_calc (ospf-packet) and checksum_calc (isis-packet) carried
two independent copies of the RFC 1008 Fletcher mod-255 adjustment math.
They are provably the same function: IS-IS is the cksum_offset=12
specialization of OSPF's cksum_offset=14 form (sop = len-13 vs
len-cksum_offset-1; guards `< 13` and `<= 12` coincide; both over
fletcher::calc_fletcher16).

Add `packet_utils::fletcher_lsa_checksum(data, cksum_offset)` (packet-utils
already depends on `fletcher`) and delegate both codecs to it: OSPF's
`lsa_checksum_calc` becomes a thin wrapper (offset 14), IS-IS's
`checksum_calc` returns `fletcher_lsa_checksum(data, 12).to_be_bytes()`.
`is_valid_checksum` stays IS-IS-local.

A known-answer test locks the IS-IS equivalence against the original
formula (kept verbatim in the test) across several lengths; the existing
OSPF (offset 14) and IS-IS round-trip/checksum tests cover the rest.
