+++ crates/packet-utils/src/lib.rs
@@ -2,6 +2,7 @@ mod admin_group;
 mod algo;
 mod bounded_capacity;
 mod checksum6;
+mod fletcher;
 mod many0;
 mod safe_split_at;
 mod sid_label;
@@ -11,6 +12,7 @@ pub use admin_group::*;
 pub use algo::*;
 pub use bounded_capacity::*;
 pub use checksum6::*;
+pub use fletcher::*;
 pub use many0::*;
 pub use safe_split_at::*;
 pub use sid_label::*;
